backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
urllib3==2.2.1
